작업사항(21/12/7) - Editor Camera(Scene 내의 자유시점 기능)
Engine Client

1. Editor
Manager
EditorObjectManager 클래스 구현
1) GameObjectEx 객체들을 초기화, 업데이트 관리

EditorObject
GameObjectEx 클래스 구현
1) Engine Library에서 사용하는 GameObject 클래스를 상속받아 필요한 부분만 오버라이딩(가상 함수 오버라이딩x)하여 사용
2) EditorObjectManager에서 관리

CameraEx 클래스 구현
1) Engine Library에서 사용하는 Camera 클래스를 상속받아 필요한 부분만 오버라이딩(가상 함수 오버라이딩x)하여 사용
2) 해당 Component를 상속받는 GameObjectEx는 EditorCamera의 기능을 사용 가능(자유시점으로 Scene 돌아다니기)
3) 자유시점으로 Scene을 돌아다닐 수 있도록 ControlEditorCamera() 함수 구현

2. TODO
GUI 배치를 Unity와 동일하게 구성하기

Engine Library

1. Manager
RenderManager 클래스 수정
1) Editor Camera 관련 멤버 변수 및 함수 추가
2) 현재 클라이언트의 상태 및 에디터 상태에 따라 렌더링 모드 변경 기능 추가 (RenderTitle()은 추후에 작업)

2. Scene
Component
Camera 클래스 수정
1)Update() 함수 삭제 -> 카메라의 자유시점 움직임은 Client의 Editor Camera가 대신함

3. TODO
Script를 어떻게 효율적으로 관리할지 고민

+++ Project/DX11_2D_GameEngine/Editor/GUI/3. Hierarchy/GUI_Hierarchy.cpp
@@ -73,7 +73,7 @@ void GUI_Hierarchy::AddGameObject(GameObject* game_object)
 	}
 
 	//노드 추가해서 화면에 보여주는 역할
-	const auto is_node_open = ImGui::TreeNodeEx(reinterpret_cast<void*>(static_cast<intptr_t>(game_object->GetObjectID())), flags, game_object->GetObjectName().c_str());
+	const auto is_node_open = ImGui::TreeNodeEx(reinterpret_cast<void*>(game_object->GetObjectID()), flags, game_object->GetObjectName().c_str());
 
 	if (m_clicked_game_object != nullptr && EditorHelper::GetInstance()->GetSelectedGameObject() != nullptr)
 	{
